Updates to fix CI
