+++ packages/evalite-tests/vitest.config.ts
@@ -3,5 +3,6 @@ import { defineConfig } from "vitest/config";
 export default defineConfig({
   test: {
     maxConcurrency: process.env.CI ? 1 : 5,
+    include: ["tests/**/*.test.ts", "!tests/fixtures/**/*.test.ts"],
   },
 });
